# wine/openapi/tml_wine_generated/openapi_server/models/token.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class Token(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, access_token=None, token_type=None):  # noqa: E501
        """Token - a model defined in OpenAPI

        :param access_token: The access_token of this Token.  # noqa: E501
        :type access_token: object
        :param token_type: The token_type of this Token.  # noqa: E501
        :type token_type: object
        """
        self.openapi_types = {
            'access_token': object,
            'token_type': object
        }

        self.attribute_map = {
            'access_token': 'access_token',
            'token_type': 'token_type'
        }

        self._access_token = access_token
        self._token_type = token_type

    @classmethod
    def from_dict(cls, dikt) -> 'Token':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Token of this Token.  # noqa: E501
        :rtype: Token
        """
        return util.deserialize_model(dikt, cls)

    @property
    def access_token(self):
        """Gets the access_token of this Token.


        :return: The access_token of this Token.
        :rtype: object
        """
        return self._access_token

    @access_token.setter
    def access_token(self, access_token):
        """Sets the access_token of this Token.


        :param access_token: The access_token of this Token.
        :type access_token: object
        """
        if access_token is None:
            raise ValueError("Invalid value for `access_token`, must not be `None`")  # noqa: E501

        self._access_token = access_token

    @property
    def token_type(self):
        """Gets the token_type of this Token.


        :return: The token_type of this Token.
        :rtype: object
        """
        return self._token_type

    @token_type.setter
    def token_type(self, token_type):
        """Sets the token_type of this Token.


        :param token_type: The token_type of this Token.
        :type token_type: object
        """
        if token_type is None:
            raise ValueError("Invalid value for `token_type`, must not be `None`")  # noqa: E501

        self._token_type = token_type