# wine/openapi/tml_wine_generated/openapi_server/models/user.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class User(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, username=None, email=None, full_name=None, disabled=None):  # noqa: E501
        """User - a model defined in OpenAPI

        :param username: The username of this User.  # noqa: E501
        :type username: object
        :param email: The email of this User.  # noqa: E501
        :type email: object
        :param full_name: The full_name of this User.  # noqa: E501
        :type full_name: object
        :param disabled: The disabled of this User.  # noqa: E501
        :type disabled: object
        """
        self.openapi_types = {
            'username': object,
            'email': object,
            'full_name': object,
            'disabled': object
        }

        self.attribute_map = {
            'username': 'username',
            'email': 'email',
            'full_name': 'full_name',
            'disabled': 'disabled'
        }

        self._username = username
        self._email = email
        self._full_name = full_name
        self._disabled = disabled

    @classmethod
    def from_dict(cls, dikt) -> 'User':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The User of this User.  # noqa: E501
        :rtype: User
        """
        return util.deserialize_model(dikt, cls)

    @property
    def username(self):
        """Gets the username of this User.


        :return: The username of this User.
        :rtype: object
        """
        return self._username

    @username.setter
    def username(self, username):
        """Sets the username of this User.


        :param username: The username of this User.
        :type username: object
        """
        if username is None:
            raise ValueError("Invalid value for `username`, must not be `None`")  # noqa: E501

        self._username = username

    @property
    def email(self):
        """Gets the email of this User.


        :return: The email of this User.
        :rtype: object
        """
        return self._email

    @email.setter
    def email(self, email):
        """Sets the email of this User.


        :param email: The email of this User.
        :type email: object
        """

        self._email = email

    @property
    def full_name(self):
        """Gets the full_name of this User.


        :return: The full_name of this User.
        :rtype: object
        """
        return self._full_name

    @full_name.setter
    def full_name(self, full_name):
        """Sets the full_name of this User.


        :param full_name: The full_name of this User.
        :type full_name: object
        """

        self._full_name = full_name

    @property
    def disabled(self):
        """Gets the disabled of this User.


        :return: The disabled of this User.
        :rtype: object
        """
        return self._disabled

    @disabled.setter
    def disabled(self, disabled):
        """Sets the disabled of this User.


        :param disabled: The disabled of this User.
        :type disabled: object
        """

        self._disabled = disabled