# wine/openapi/tml_wine_generated/openapi_server/models/validation_error.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class ValidationError(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, loc=None, msg=None, type=None):  # noqa: E501
        """ValidationError - a model defined in OpenAPI

        :param loc: The loc of this ValidationError.  # noqa: E501
        :type loc: object
        :param msg: The msg of this ValidationError.  # noqa: E501
        :type msg: object
        :param type: The type of this ValidationError.  # noqa: E501
        :type type: object
        """
        self.openapi_types = {
            'loc': object,
            'msg': object,
            'type': object
        }

        self.attribute_map = {
            'loc': 'loc',
            'msg': 'msg',
            'type': 'type'
        }

        self._loc = loc
        self._msg = msg
        self._type = type

    @classmethod
    def from_dict(cls, dikt) -> 'ValidationError':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ValidationError of this ValidationError.  # noqa: E501
        :rtype: ValidationError
        """
        return util.deserialize_model(dikt, cls)

    @property
    def loc(self):
        """Gets the loc of this ValidationError.


        :return: The loc of this ValidationError.
        :rtype: object
        """
        return self._loc

    @loc.setter
    def loc(self, loc):
        """Sets the loc of this ValidationError.


        :param loc: The loc of this ValidationError.
        :type loc: object
        """
        if loc is None:
            raise ValueError("Invalid value for `loc`, must not be `None`")  # noqa: E501

        self._loc = loc

    @property
    def msg(self):
        """Gets the msg of this ValidationError.


        :return: The msg of this ValidationError.
        :rtype: object
        """
        return self._msg

    @msg.setter
    def msg(self, msg):
        """Sets the msg of this ValidationError.


        :param msg: The msg of this ValidationError.
        :type msg: object
        """
        if msg is None:
            raise ValueError("Invalid value for `msg`, must not be `None`")  # noqa: E501

        self._msg = msg

    @property
    def type(self):
        """Gets the type of this ValidationError.


        :return: The type of this ValidationError.
        :rtype: object
        """
        return self._type

    @type.setter
    def type(self, type):
        """Sets the type of this ValidationError.


        :param type: The type of this ValidationError.
        :type type: object
        """
        if type is None:
            raise ValueError("Invalid value for `type`, must not be `None`")  # noqa: E501

        self._type = type